+++ modal/functions.py
@@ -350,6 +350,8 @@ class _Function(typing.Generic[P, T], _Object, type_prefix="fu"):
                 is_method=True,
                 use_function_id=class_service_function.object_id,
                 use_method_name=method_name,
+                batch_max_size=partial_function.batch_max_size or 0,
+                batch_linger_ms=partial_function.batch_wait_ms or 0,
             )
             assert resolver.app_id
             request = api_pb2.FunctionCreateRequest(
@@ -491,6 +493,8 @@ class _Function(typing.Generic[P, T], _Object, type_prefix="fu"):
         timeout: Optional[int] = None,
         concurrency_limit: Optional[int] = None,
         allow_concurrent_inputs: Optional[int] = None,
+        batch_max_size: Optional[int] = None,
+        batch_wait_ms: Optional[int] = None,
         container_idle_timeout: Optional[int] = None,
         cpu: Optional[float] = None,
         keep_warm: Optional[int] = None,  # keep_warm=True is equivalent to keep_warm=1
@@ -660,6 +664,14 @@ class _Function(typing.Generic[P, T], _Object, type_prefix="fu"):
             else:
                 raise InvalidError("Webhooks cannot be generators")
 
+        if info.raw_f and batch_max_size:
+            func_name = info.raw_f.__name__
+            if is_generator:
+                raise InvalidError(f"Modal batched function {func_name} cannot return generators")
+            for arg in inspect.signature(info.raw_f).parameters.values():
+                if arg.default is not inspect.Parameter.empty:
+                    raise InvalidError(f"Modal batched function {func_name} does not accept default arguments.")
+
         if container_idle_timeout is not None and container_idle_timeout <= 0:
             raise InvalidError("`container_idle_timeout` must be > 0")
 
@@ -815,6 +827,8 @@ class _Function(typing.Generic[P, T], _Object, type_prefix="fu"):
                     app_name=app_name,
                     is_builder_function=is_builder_function,
                     allow_concurrent_inputs=allow_concurrent_inputs or 0,
+                    batch_max_size=batch_max_size or 0,
+                    batch_linger_ms=batch_wait_ms or 0,
                     worker_id=config.get("worker_id"),
                     is_auto_snapshot=is_auto_snapshot,
                     is_method=bool(info.user_cls) and not info.is_service_class(),